# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.GetRoomAvailabilityRequest:
    properties:
      end_date:
        type: string
      start_date:
        type: string
    required:
    - end_date
    - start_date
    type: object
  handlers.GetRoomStatusRequest:
    properties:
      end_date:
        type: string
      start_date:
        type: string
    required:
    - end_date
    - start_date
    type: object
...
  models.Booking:
    properties:
      booking_date:
        type: string
      booking_id:
        type: integer
      booking_name:
        type: string
      check_in_date:
        type: string
      check_out_date:
        type: string
      receipt:
        $ref: '#/definitions/models.Receipt'
      room:
        $ref: '#/definitions/models.Room'
      room_num:
        type: integer
      statuses:
        items:
          $ref: '#/definitions/models.RoomStatus'
        type: array
      total_price:
        type: integer
    required:
    - booking_name
    - check_in_date
    - check_out_date
    - room_num
    type: object
  models.Receipt:
    properties:
      amount:
        type: integer
      booking_id:
        type: integer
      issue_date:
        type: string
      payment_date:
        type: string
      payment_method:
        enum:
        - Credit
        - Debit
        - Bank
        - Transfer
        type: string
      receipt_id:
        type: integer
    required:
    - amount
    - payment_method
    type: object
  models.Room:
    properties:
      bookings:
        items:
          $ref: '#/definitions/models.Booking'
        type: array
      room_num:
        type: integer
      room_type:
        $ref: '#/definitions/models.RoomType'
      statuses:
        items:
          $ref: '#/definitions/models.RoomStatus'
        type: array
      type_id:
        type: integer
    type: object
  models.RoomFacility:
    properties:
      fac_id:
        type: integer
      type_id:
        type: integer
    type: object
  models.RoomStatus:
    properties:
      booking:
        $ref: '#/definitions/models.Booking'
      booking_id:
        type: integer
      calendar:
        type: string
      room_num:
        type: integer
      status:
        enum:
        - Available
        - Occupied
        type: string
    required:
    - status
    type: object
  models.RoomType:
    properties:
      area:
        type: integer
      description:
        type: string
      name:
        type: string
      noOfGuest:
        type: integer
      price_per_night:
        type: integer
      room_facilities:
        items:
          $ref: '#/definitions/models.RoomFacility'
        type: array
      rooms:
        items:
          $ref: '#/definitions/models.Room'
        type: array
      type_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for hotel booking system with transaction management and concurrency
    control
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Hotel Booking System API
  version: "1.0"
paths:
...
  /rooms:
    get:
      description: Get all rooms with their types
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Room'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all rooms
      tags:
      - rooms
  /rooms/{roomNum}:
    get:
      description: Get a room by its room number
      parameters:
      - description: Room Number
        in: path
        name: roomNum
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Room'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a room by number
      tags:
      - rooms
  /rooms/{roomNum}/status:
    post:
      consumes:
      - application/json
      description: Get room status for a specific room and date range
      parameters:
      - description: Room Number
        in: path
        name: roomNum
        required: true
        type: integer
      - description: Date range
        in: body
        name: dates
        required: true
        schema:
          $ref: '#/definitions/handlers.GetRoomStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.RoomStatus'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get room status
      tags:
      - rooms
  /rooms/availability:
    post:
      consumes:
      - application/json
      description: Get availability summary for all rooms in a date range
      parameters:
      - description: Date range
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.GetRoomAvailabilityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get room availability summary
      tags:
      - rooms
...
  /rooms/type/{typeId}:
    get:
      description: Get all rooms of a specific type
      parameters:
      - description: Room Type ID
        in: path
        name: typeId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Room'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get rooms by type
      tags:
      - rooms
  /rooms/types:
    get:
      description: Get all available room types
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.RoomType'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all room types
      tags:
      - rooms
swagger: "2.0"